pyproject.toml
[project]
name = "sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py==2.3.2",
    "fastapi[standard]>=0.135.2",
    "feedparser==6.0.10",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.42.0",
]
